Moved data.py to relevant worker and updated file path
